Update Bijective method
The bijective method compares two lists of files stored in two "watchdog" files, and compares each item according to the names. 
The outputs are two files "list1.txt" and "list2.txt" that contain all possible relations between the files.
The bijective method does not produce a real bijective relation as two files can be significantly close but not in the same place on disk.
So a single file can be comparable to two others that are not located in the same place.
"list1.txt" and "list2.txt" can be compared using another method to compare the files more deeply, like byte per byte or per value according to NPZ files.